requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pinecone-client
pymongo
groq
pydub
speechrecognition
gtts
elevenlabs
streamlit-webrtc
aiofiles
streamlit-audiorec  